+++ src/lib/layout/search/SearchResultItem.tsx
@@ -1,66 +1,30 @@
-import { Flex, Image } from "@chakra-ui/react";
+import { Flex } from "@chakra-ui/react";
 
-import type { IconKeys } from "lib/components/icon";
-import { CustomIcon } from "lib/components/icon";
-import type {
-  ResultMetadata,
-  SearchResultType,
-} from "lib/services/searchService";
+import type { SearchResult } from "lib/services/searchService";
 import type { Option } from "lib/types";
 
 import { SearchResultItemBody } from "./SearchResultItemBody";
+import { SearchResultItemIcon } from "./SearchResultItemIcon";
 import { getRouteOptions } from "./utils";
 
 interface ResultItemProps {
   index: number;
-  type: SearchResultType;
-  value: string;
+  result: SearchResult;
   cursor: Option<number>;
-  metadata: ResultMetadata;
   setCursor: (index: Option<number>) => void;
-  handleSelectResult: (type?: SearchResultType, isClick?: boolean) => void;
-  onClose?: () => void;
+  handleSelectResult: (result?: SearchResult, isClick?: boolean) => void;
 }
-
-const getIcon = (type: Option<SearchResultType>) => {
-  switch (type) {
-    case "Account Address":
-      return "admin" as IconKeys;
-    case "Transaction Hash":
-      return "file" as IconKeys;
-    case "Code ID":
-      return "code" as IconKeys;
-    case "Contract Address":
-    case "Module Path":
-      return "contract-address" as IconKeys;
-    case "Block":
-      return "block" as IconKeys;
-    case "Proposal ID":
-      return "proposal" as IconKeys;
-    case "Validator Address":
-      return "validator" as IconKeys;
-    case "Pool ID":
-      return "pool" as IconKeys;
-    case "NFT Collection Address":
-      return "group" as IconKeys;
-    default:
-      return "list" as IconKeys;
-  }
-};
 
 export const SearchResultItem = ({
   index,
-  type,
-  value,
+  result,
   cursor,
-  metadata,
   setCursor,
   handleSelectResult,
-  onClose,
 }: ResultItemProps) => {
-  const route = getRouteOptions(type)?.pathname;
+  const route = getRouteOptions(result.type)?.pathname;
   const isAccountAddress =
-    type === "Account Address" || type === "Contract Address";
+    result.type === "Account Address" || result.type === "Contract Address";
 
   return (
     <Flex id={`item-${index}`}>
@@ -76,25 +40,17 @@ export const SearchResultItem = ({
           transition="all 0.25s ease-in-out"
           bg={index === cursor ? "gray.700" : undefined}
           onMouseMove={() => index !== cursor && setCursor(index)}
-          onClick={() => {
-            handleSelectResult(type, true);
-            onClose?.();
-          }}
+          onClick={() => handleSelectResult(result, true)}
         >
-          {isAccountAddress && metadata.initiaUsername?.username ? (
-            <Image
-              src="https://assets.alleslabs.dev/webapp-assets/name-services/initia-username.svg"
-              borderRadius="full"
-              width={5}
-              height={5}
-            />
-          ) : (
-            <CustomIcon name={getIcon(type)} color="gray.600" />
-          )}
+          <SearchResultItemIcon
+            type={result.type}
+            isInitiaUsername={
+              isAccountAddress && !!result.metadata?.initiaUsername
+            }
+            isIcns={isAccountAddress && !!result.metadata?.icns}
+          />
           <SearchResultItemBody
-            metadata={metadata}
-            value={value}
-            type={type}
+            result={result}
             isAccountAddress={isAccountAddress}
           />
         </Flex>