+++ src/lib/services/nft-collection/lcd.ts
@@ -1,18 +1,18 @@
 /* eslint-disable sonarjs/no-duplicate-string */
 import { z } from "zod";
 
-import { getMoveViewJson } from "../move/module/api";
+import { getMoveViewJsonLcd } from "../move/module/lcd";
 import type { CollectionByCollectionAddressResponse } from "../types";
 import { zHexAddr } from "lib/types";
 import type { HexAddr, HexAddr32 } from "lib/types";
 import { parseWithError } from "lib/utils";
 
-export const getCollectionByCollectionAddressLite = async (
+export const getCollectionByCollectionAddressLcd = async (
   endpoint: string,
   collectionAddress: HexAddr32
 ) =>
   Promise.all([
-    getMoveViewJson(
+    getMoveViewJsonLcd(
       endpoint,
       "0x1" as HexAddr,
       "collection",
@@ -20,7 +20,7 @@ export const getCollectionByCollectionAddressLite = async (
       ["0x1::collection::Collection"],
       [`"${collectionAddress}"`]
     ).then((name) => parseWithError(z.string(), name)),
-    getMoveViewJson(
+    getMoveViewJsonLcd(
       endpoint,
       "0x1" as HexAddr,
       "collection",
@@ -28,7 +28,7 @@ export const getCollectionByCollectionAddressLite = async (
       ["0x1::collection::Collection"],
       [`"${collectionAddress}"`]
     ).then((name) => parseWithError(z.string(), name)),
-    getMoveViewJson(
+    getMoveViewJsonLcd(
       endpoint,
       "0x1" as HexAddr,
       "collection",
@@ -36,7 +36,7 @@ export const getCollectionByCollectionAddressLite = async (
       ["0x1::collection::Collection"],
       [`"${collectionAddress}"`]
     ).then((name) => parseWithError(z.string(), name)),
-    getMoveViewJson(
+    getMoveViewJsonLcd(
       endpoint,
       "0x1" as HexAddr,
       "collection",